pt-not-main is nothing
I created an extra folder and send the unused, test js files and other files that are unrelated to the project. 'Cuz I thought I was getting errors because of them.

I think I can't do run method to the code file that uses puppeteer, since in the puppeteer tutorial video, to work the puppeteer code, it use node cleaner-index.js, i think I'm getting errors because of that.