downloading the generated element
Making the template done, downloading the element left.
Turns out best way to download the image is also to use puppeteer's screenshot property and it can screenshot specific element too.

the problem is after creating the element, i don't know how to send the element to take the screenshot of the element.

and for extracting the review there are things need to be done to extract all of the review and in the template i need to put spoiler alert.