Everything finished
tells if there are spoilers. downloading all done.

what's left are styling, and if possible, it'd be better to not wait for all the elements to be finished loading and after all the required elements that needs to be scraped are done loaded then stopping would save a lot more time, in my opinion.